refactor Login component and make input errors show properly
