+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Route } from "react-router-dom";
 import Onboarding from "./components/Onboarding";
 import "./App.css";
-import PrivateRoute from "./components/PrivateRoute";
+import PrivateRoute from "./components/common/PrivateRoute";
 import Login from "./components/auth/LogIn";
 import Dashboard from "./components/Dashboard";
 import SignUp from "./components/auth/SignUp";
