+++ src/components/common/PrivateRoute.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
-import { isLoggedIn } from "../utils";
+import { isLoggedIn } from "../../utils";
 
 const PrivateRoute = ({ component: Component, ...rest }) => {
   const isSignedIn = isLoggedIn();
